move `jwt_utils.py` to `utils/` dir, add SECRET_KEY placeholder to .env.template
