+++ backend/.env.template
@@ -1,1 +1,2 @@
 DATABASE_URL=mysql+pymysql://root@localhost:3306/bp
+SECRET_KEY=put_secret_key_here
